Переписаны тесты для проверки исключений
