+++ src/test/java/praktikum/IngredientTest.java
@@ -13,23 +13,59 @@ public class IngredientTest {
         ingredient = new Ingredient(IngredientType.SAUCE,"name", 10);
     }
     @Test
-    public void testIngredientgetPrice() {
+    public void testIngredientGetPrice() {
         assertEquals(10, ingredient.getPrice(),0);
     }
 
     @Test
-    public void getName() {
+    public void testIngredientGetName() {
         assertEquals("name", ingredient.getName());
     }
 
     @Test
-    public void getTypeSauce() {
+    public void testIngredientGetTypeSauce() {
         assertEquals(IngredientType.SAUCE, ingredient.getType());
     }
-    @Test
-    public void TypeIsNullValue() {
-        ingredient = new Ingredient(null,"name", 10);
+
+    @Test(expected = Exception.class)
+    public void testIngredientTypeIsNullValueExpectedException() {
+        try {
+            ingredient = new Ingredient(null,"name", 10);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
 
     }
+
+    @Test (expected = Exception.class)
+    public void testNameIsNullValueExpectedException() {
+        try {
+            ingredient = new Ingredient(IngredientType.SAUCE,null, 10);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+    }
+    @Test (expected = Exception.class)
+    public void testNameIsEmptyValueExpectedException()  throws  Exception {
+        try {
+            ingredient = new Ingredient(IngredientType.SAUCE,"", 10);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+    }
+
+    @Test (expected = Exception.class)
+    public void testPriceIsNegativeExpectedException()  throws  Exception {
+        try {
+            ingredient = new Ingredient(IngredientType.SAUCE,"name", -10);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+    }
+
+
 
 }
